Convert bloom and tonemapping to graphics features
